+++ GHEtool/gui/data_2_borefield_func.py
@@ -6,6 +6,7 @@ from __future__ import annotations
 from functools import partial
 from typing import TYPE_CHECKING
 import logging
+import warnings
 
 import numpy as np
 from GHEtool import Borefield, FluidData, MultipleUTube, GroundConstantTemperature, GroundFluxTemperature, GroundTemperatureGradient, CoaxialPipe
@@ -93,11 +94,14 @@ def data_2_borefield(ds: DataStorage) -> tuple[Borefield, partial[[], None]]:
             hourly_data.hourly_cooling_load = peak_cooling * (1 + 1 / ds.SEER)
             hourly_data.hourly_heating_load = peak_heating * (1 - 1 / ds.SCOP)
             borefield.load = hourly_data
+            borefield._secundary_borefield_load.hourly_heating_load = peak_heating
+            borefield._secundary_borefield_load.hourly_cooling_load = peak_cooling
 
     # add dhw when needed
     if ds.option_include_dhw == 1 and not ds.aim_optimize:
         SCOP = ds.SCOP_DHW if ds.geo_load == 1 else 99999999999
         borefield.load.dhw = ds.DHW * (1 - 1 / SCOP)
+        borefield._secundary_borefield_load.dhw = ds.DHW
 
     # set up the borefield sizing
     borefield.calculation_setup(use_constant_Rb=ds.option_method_rb_calc == 0,
@@ -202,6 +206,10 @@ def _create_fluid_data(ds: DataStorage) -> FluidData:
     """
     fluid_data = FluidData(ds.option_fluid_mass_flow, ds.option_fluid_conductivity, ds.option_fluid_density, ds.option_fluid_capacity, ds.option_fluid_viscosity)
     if ds.option_fluid_selector == 1:
+
+        # set warnings as error
+        warnings.filterwarnings("error")
+
         # use pygfunction to get fluid properties
         if ds.option_glycol_selector == 0:
             fluid_data.import_fluid_from_pygfunction(gt.media.Fluid('MEG',
@@ -209,6 +217,7 @@ def _create_fluid_data(ds: DataStorage) -> FluidData:
         else:
             fluid_data.import_fluid_from_pygfunction(gt.media.Fluid('MPG',
                                                                     ds.option_glycol_percentage, ds.option_fluid_ref_temp))
+    warnings.resetwarnings()
     return fluid_data