+++ CHANGELOG.md
@@ -3,6 +3,20 @@ All notable changes to this project will be documented in this file. For future 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.2.1] - 2024
+
+## Added
+- Equal functions for HourlyGeothermalLoad and MonthlyGeothermalLoadAbsolute (issue #189).
+
+## changed
+- Negative reference temperatures for the fluid are now possible (issue #192).
+
+## fixed
+- Bug in load-duration curve when not working with optimize load profile (issue #189).
+- Bug in hourly data (issue #196).
+- Bug in saving after a file has been moved (issue #198).
+
+
 ## [2.2.0] - 2023-10-17
 
 ## Added